requirements.txt
python-telegram-bot==23.1
gspread>=6.0.0
oauth2client>=4.1.3
pytz
